+++ web/app/blog/page.tsx
@@ -1,0 +1,38 @@
+import type { NextPage } from "next";
+import Link from "next/link";
+
+import { articlesFrontmatters } from "@common/mdxLoader";
+import BlogPostCard from "@components/BlogPostCard";
+import Pagination from "@components/Pagination";
+
+export const metadata = {
+  title: "Blog",
+};
+
+export const dynamic = "force-static";
+
+const BlogListing: NextPage = async () => {
+  const frontmatters = await articlesFrontmatters();
+  console.log(frontmatters);
+
+  const postsCards = frontmatters.map((post) => (
+    <li key={post.slug}>
+      <BlogPostCard post={post} />
+    </li>
+  ));
+
+  return (
+    <>
+      <h1 className="heading-primary">The Blog</h1>
+      <p className="mb-6">
+        Or check out blogposts sorted by{" "}
+        <Link href={"/series"} className="link">
+          series
+        </Link>
+      </p>
+      <ul className="flex flex-col divide-y">{postsCards}</ul>
+    </>
+  );
+};
+
+export default BlogListing;
