+++ web/common/components/BlogPostCard.tsx
@@ -1,6 +1,4 @@
 import type { ArticleFrontmatter } from "@common/mdxLoader";
-import type * as Schema from "@common/sanityTypes";
-import { getBeginningOfArticle, isPartOfSeries } from "@common/utils/blogpost";
 import { formatDate } from "@common/utils/misc";
 import Link from "next/link";
 import { Pills } from "./Pill";
@@ -10,9 +8,7 @@ interface BlogPostCardProps {
 }
 
 export default function BlogPostCard({ post }: BlogPostCardProps) {
-  const tags = post.tags || [];
-
-  const date = new Date(post.releaseDate);
+  const date = post.releaseDate;
   const formattedDate = formatDate(date);
 
   // const truncatedText = getBeginningOfArticle(post, 250);
@@ -26,10 +22,8 @@ export default function BlogPostCard({ post }: BlogPostCardProps) {
         <div className="flex gap-4 text-md pb-3 font-semibold">
           <span className="">{formattedDate}</span>
         </div>
-        <p className="two-line-text-ellipsis text-sm h-10 m-0">
-          todo
-        </p>
-        <Pills texts={tags.map(({ label }) => label)} />
+        <p className="two-line-text-ellipsis text-sm h-10 m-0">todo</p>
+        <Pills texts={post.tags.map(({ label }) => label)} />
       </div>
     </Link>
   );
